add soup list component
